Split di factories by modules
