+++ workers/loc.api/di/factories/data-inserter-factory.js
@@ -1,0 +1,14 @@
+'use strict'
+
+const TYPES = require('../types')
+
+module.exports = (ctx) => {
+  return (syncColls) => {
+    const dataInserter = ctx.container.get(
+      TYPES.DataInserter
+    )
+    dataInserter.init(syncColls)
+
+    return dataInserter
+  }
+}
