+++ workers/loc.api/di/factories/db-migrator-factory.js
@@ -1,0 +1,24 @@
+'use strict'
+
+const TYPES = require('../types')
+
+module.exports = (ctx) => {
+  const { dbDriver } = ctx.container.get(
+    TYPES.CONF
+  )
+
+  return () => {
+    const dao = ctx.container.get(
+      TYPES.DAO
+    )
+
+    if (dbDriver === 'sqlite') {
+      const sqliteDbMigrator = ctx.container.get(
+        TYPES.SqliteDbMigrator
+      )
+      sqliteDbMigrator.setDao(dao)
+
+      return sqliteDbMigrator
+    }
+  }
+}
